requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
motor==3.3.2
python-dotenv==1.0.1
google-generativeai==0.3.2
beautifulsoup4==4.12.3
requests==2.31.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2